+++ autotransform/validator/script.py
@@ -7,37 +7,43 @@
 
 # @black_format
 
-"""The implementation for the ScriptTransformer."""
+"""The implementation for the ScriptValidator."""
 
 from __future__ import annotations
 
 import json
 import subprocess
 from tempfile import NamedTemporaryFile as TmpFile
-from typing import Any, List, Mapping, Optional, TypedDict
+from typing import Any, List, Mapping, Optional, Sequence, TypedDict
 
 from typing_extensions import NotRequired
 
+import autotransform.schema
 from autotransform.batcher.base import Batch
 from autotransform.event.debug import DebugEvent
 from autotransform.event.handler import EventHandler
 from autotransform.item.base import Item
-from autotransform.transformer.base import Transformer
-from autotransform.transformer.type import TransformerType
+from autotransform.item.file import FileItem
+from autotransform.validator.base import ValidationResult, ValidationResultLevel, Validator
+from autotransform.validator.type import ValidatorType
 
 
-class ScriptTransformerParams(TypedDict):
-    """The param type for a ScriptTransformer."""
+class ScriptValidatorParams(TypedDict):
+    """The param type for a ScriptValidator"""
 
     script: str
     args: List[str]
-    timeout: int
+    failure_level: ValidationResultLevel
     per_item: NotRequired[bool]
+    run_on_changes: NotRequired[bool]
 
 
-class ScriptTransformer(Transformer[ScriptTransformerParams]):
-    """A Transformer that makes changes using an invoked script. Sentinel values can be used in the
-    args to supply information from the Batch.
+class ScriptValidator(Validator[ScriptValidatorParams]):
+    """Runs a script with the supplied arguments to perform validation. If the per_item flag is
+    set to True, the script will be invoked on each Item. If run_on_changes is set to True, the
+    script will replace the Batch items with FileItems for each changed file. The failure_level
+    indicates the result level if the script returns a non-zero exit code. Sentinel values can
+    be used in args to provide custom arguments for a run.
     The available sentinel values for args are:
         <<KEY>>: A json encoded list of the input files for a batch. If the per_item flag is set in
             params, this will simply be the key of an Item.
@@ -49,41 +55,62 @@ class ScriptTransformer(Transformer[ScriptTransformerParams]):
      with a path to a file containing the value.
 
     Attributes:
-        _params (ScriptTransformerParams): The script and args, along with the timeout and a
-            per_item flag for invoking the script on each Item.
+        _params (ScriptValidatorParams): Contains the args and set-up for the script.
     """
 
-    _params: ScriptTransformerParams
+    _params: ScriptValidatorParams
 
     @staticmethod
-    def get_type() -> TransformerType:
-        """Used to map Transformer components 1:1 with an enum, allowing construction from JSON.
+    def get_type() -> ValidatorType:
+        """Used to map Validator components 1:1 with an enum, allowing construction from JSON.
 
         Returns:
-            TransformerType: The unique type associated with this Transformer.
+            ValidatorType: The unique type associated with this Validator.
         """
 
-        return TransformerType.SCRIPT
+        return ValidatorType.SCRIPT
 
-    def transform(self, batch: Batch) -> None:
-        """Runs the script transformation against the Batch, either on each item individually or
-        on the entire Batch, based on the per_item flag.
+    def validate(self, batch: Batch) -> ValidationResult:
+        """Runs the script validation against the Batch, either on each item individually or
+        on the entire Batch, based on the per_item flag. If the script returns a non-zero exit
+        code, the failure_level in params will be in the result.
 
         Args:
-            batch (Batch): The Batch being transformed.
+            batch (Batch): The transformed Batch to validate.
+
+        Returns:
+            ValidationResult: The result of the validation check indicating the severity of any
+                validation failures as well as an associated message
         """
 
         if self._params.get("per_item", False):
-            for item in batch["items"]:
-                self._transform_single(item, batch.get("metadata", None))
-        else:
-            self._transform_batch(batch)
+            if self._params.get("run_on_changes", False):
+                current_schema = autotransform.schema.current
+                assert current_schema is not None
+                repo = current_schema.get_repo()
+                assert repo is not None
+                items: Sequence[Item] = [FileItem(file) for file in repo.get_changed_files(batch)]
+            else:
+                items = batch["items"]
 
-    def _transform_single(self, item: Item, batch_metadata: Optional[Mapping[str, Any]]) -> None:
-        """Executes a simple script to transform a single Item. Sentinel values can be used
+            for item in items:
+                result = self._validate_single(item, batch.get("metadata", None))
+                if result["level"] != ValidationResultLevel.NONE:
+                    return result
+            return {
+                "level": ValidationResultLevel.NONE,
+                "message": "",
+                "validator": self.get_type(),
+            }
+        return self._validate_batch(batch)
+
+    def _validate_single(
+        self, item: Item, batch_metadata: Optional[Mapping[str, Any]]
+    ) -> ValidationResult:
+        """Executes a simple script to validate a single Item. Sentinel values can be used
         in args that will be replaced when the script is invoked.
         The available sentinel values for args are:
-            <<KEY>>: The key of the Item being transformed.
+            <<KEY>>: The key of the Item being validated.
             <<EXTRA_DATA>>: A JSON encoding of the Item's extra_data. If extra_data is not present,
                 it is treated as an empty Dict.
             <<METADATA>>: A JSON encoded version of the Batch's metadata.
@@ -91,7 +118,7 @@ class ScriptTransformer(Transformer[ScriptTransformerParams]):
             replaced with a path to a file containing the value.
 
         Args:
-            item (Item): The Item that will be transformed.
+            item (Item): The Item that will be validated.
             batch_metadata (Optional[Mapping[str, Any]]): The metadata of the Batch containing the
                 Item.
         """
@@ -103,12 +130,11 @@ class ScriptTransformer(Transformer[ScriptTransformerParams]):
         extra_data = item.get_extra_data()
         if extra_data is None:
             extra_data = {}
-        metadata = batch_metadata if batch_metadata is not None else {}
 
         arg_replacements = {
             "<<KEY>>": item.get_key(),
             "<<EXTRA_DATA>>": json.dumps(extra_data),
-            "<<METADATA>>": json.dumps(metadata),
+            "<<METADATA>>": json.dumps(batch_metadata),
         }
 
         with TmpFile(mode="w+") as inp, TmpFile(mode="w+") as meta, TmpFile(mode="w+") as extra:
@@ -123,7 +149,7 @@ class ScriptTransformer(Transformer[ScriptTransformerParams]):
             arg_replacements["<<EXTRA_DATA_FILE>>"] = extra.name
 
             # Make metadata file
-            json.dump(metadata, meta)
+            json.dump(batch_metadata, meta)
             meta.flush()
             arg_replacements["<<METADATA_FILE>>"] = meta.name
 
@@ -134,12 +160,20 @@ class ScriptTransformer(Transformer[ScriptTransformerParams]):
                 else:
                     cmd.append(arg)
 
-            # Run Script
+            # Run script
             event_handler.handle(DebugEvent({"message": f"Running command: {str(cmd)}"}))
-            subprocess.check_output(cmd, timeout=self._params["timeout"])
+            proc = subprocess.run(cmd, capture_output=True, encoding="ascii", check=False)
+        level = (
+            self._params["failure_level"] if proc.returncode != 0 else ValidationResultLevel.NONE
+        )
+        return {
+            "level": level,
+            "message": f"[{self._params['script']}] {proc.stderr}",
+            "validator": self.get_type(),
+        }
 
-    def _transform_batch(self, batch: Batch) -> None:
-        """Executes a simple script to transform the given Batch. Sentinel values can be used
+    def _validate_batch(self, batch: Batch) -> ValidationResult:
+        """Executes a simple script to validate the given Batch. Sentinel values can be used
         in args that will be replaced when the script is invoked.
         The available sentinel values for args are:
             <<KEY>>: A json encoded list of the input files for a batch.
@@ -156,11 +190,19 @@ class ScriptTransformer(Transformer[ScriptTransformerParams]):
         event_handler = EventHandler.get()
 
         cmd = [self._params["script"]]
+        if self._params.get("run_on_changes", False):
+            current_schema = autotransform.schema.current
+            assert current_schema is not None
+            repo = current_schema.get_repo()
+            assert repo is not None
+            items: Sequence[Item] = [FileItem(file) for file in repo.get_changed_files(batch)]
+        else:
+            items = batch["items"]
 
-        item_keys = [item.get_key() for item in batch["items"]]
+        item_keys = [item.get_key() for item in items]
         extra_data = {
             item.get_key(): item.get_extra_data()
-            for item in batch["items"]
+            for item in items
             if item.get_extra_data() is not None
         }
         metadata = batch.get("metadata", {})
@@ -195,32 +237,50 @@ class ScriptTransformer(Transformer[ScriptTransformerParams]):
 
             # Run script
             event_handler.handle(DebugEvent({"message": f"Running command: {str(cmd)}"}))
-            subprocess.check_output(cmd, timeout=self._params["timeout"])
+            proc = subprocess.run(cmd, capture_output=True, encoding="ascii", check=False)
+        level = (
+            self._params["failure_level"] if proc.returncode != 0 else ValidationResultLevel.NONE
+        )
+        return {
+            "level": level,
+            "message": f"[{self._params['script']}] {proc.stderr}",
+            "validator": self.get_type(),
+        }
 
     @staticmethod
-    def from_data(data: Mapping[str, Any]) -> ScriptTransformer:
-        """Produces a ScriptTransformer from the provided data.
+    def from_data(data: Mapping[str, Any]) -> ScriptValidator:
+        """Produces an instance of the component from decoded params. Implementations should
+        assert that the data provided matches expected types and is valid.
 
         Args:
             data (Mapping[str, Any]): The JSON decoded params from an encoded bundle.
 
         Returns:
-            ScriptTransformer: An instance of the ScriptTransformer.
+            ScriptValidator: An instance of the ScriptValidator.
         """
 
         script = data["script"]
         assert isinstance(script, str)
         args = data["args"]
         assert isinstance(args, List)
-        args = [str(arg) for arg in args]
-        timeout = data["timeout"]
-        assert isinstance(timeout, int)
-
-        params: ScriptTransformerParams = {"script": script, "args": args, "timeout": timeout}
+        for arg in args:
+            assert isinstance(arg, str)
+        failure_level = data["failure_level"]
+        assert isinstance(failure_level, int)
+        params: ScriptValidatorParams = {
+            "script": script,
+            "args": args,
+            "failure_level": ValidationResultLevel.from_value(failure_level),  # type: ignore
+        }
 
         per_item = data.get("per_item", None)
         if per_item is not None:
             assert isinstance(per_item, bool)
             params["per_item"] = per_item
 
-        return ScriptTransformer(params)
+        run_on_changes = data.get("run_on_changes", None)
+        if run_on_changes is not None:
+            assert isinstance(run_on_changes, bool)
+            params["run_on_changes"] = run_on_changes
+
+        return ScriptValidator(params)