+++ autotransform/validator/type.py
@@ -15,5 +15,4 @@ from enum import Enum
 class ValidatorType(str, Enum):
     """A simple enum for 1:1 Validator to type mapping."""
 
-    MYPY = "mypy"
-    PYTEST = "pytest"
+    SCRIPT = "script"