+++ CHANGELOG.md
@@ -14,8 +14,7 @@ Version 0.2.0 represents a production ready beta of AutoTransform. It includes a
  - RegexFilter as a Filter which checks the Item's key against a regex pattern
  - FileContentRegexFilter as a Filter that checks a file's contents against a regex pattern
  - Added ShardFilter concept and KeyHashShardFilter
- - MypyValidator for validating using mypy
- - PytestValidator for validating using pytest
+ - ScriptValidator for handling validation using command line scripts like mypy or pytest
 
  - Runner class added to provide an API for invoking runs
  - GithubRunner class added to use Github workflows for remote runs